Fixed JSHint errors and Boostrap JS issues
Now that we are using Bootstrap JS (not just Bootstrap CSS), some
of the components started functioning differently and appeared to
be broken. This was because we were only using half of the Boostrap
framework. Now that Bootstrap JS is included as well, I fixed the
components that broke as a result.
Also fixed various JSHint compiler errors.