Change CarePlan route loading to use promises
Returns promises for *all* queries that are being executed.
Prevents the route from loading and drawing the template before
the models are available.
This was needed to fix an issue where navigating to one Care Plan
after another would periodically cause the page to freeze because
data was changing after the template was drawn.