Updated/fixed nomination processing (change-request component)
Now, multiple nominations for a given author/resource can exist.
Each nomination for a "create" or "delete" action includes a "diff"
attribute that is equal to {}.
Each nomination for an "update" action includes a single "diff"
attribute with one path.
Accepted nominations are now applied and removed on the server
side. In addition, they are hidden from the template when accepted.