Fixed fhir-timing and fhir-datetime issues
Attributes in the fhir-timing component did not save correctly
because the intermediary models were not being created.
I changed backbone-element to make it more flexible (now it can be
used for an array attribute or a single attribute). This is fully
generic and can replace the more specific codeable-concept
component.
I change fhir-datetime to allow it to set a single Date or an array
that contains a single date.
The timing model in the Ember FHIR adapter was erroneous, had an
attribute as a date instead of a date array. I created a transform
to fix this for now, and will separately make a pull request to the
Ember FHIR adapter repo.
Modals for fhir-timing and fhir-datetime were conflicting because
each modal needs its own ID; I fixed this by using the component's
elementId (which is already unique).
The timing-repeat-component model was throwing errors, I made the
validation more robust to prevent the errors.