Fix routing, refactor directory structure, add scrollbars to pages
The old method of routing caused additional information from parent
routes to be rendered when viewing child route pages.
Flattened out logical routing so that this doesn't happen anymore.
Since routing was changed, directory structure needed to be
refactored to match.
In addition, this now lets us add scrollbars to the 'medentry' and
'medrec' pages.
Added scrollbars to medentry page; medrec page needs to be
reworked in a separate commit.