Revamped display of 'new' annotations
'New' annotations (e.g. those that the user has not seen) are
displayed when a resource is expanded, and they are immediately
marked as 'seen' but they still show until the page is refreshed.
In addition, annotations that are created by the user also show
in the expanded resource until the page is refreshed.