Fixed resource 'relatedX' attributes
The 'relatedX' computed properties now only return resources that
have not been deleted. Due to Ember limitations and the way we have
to handle nominations, when a resource is deleted, it does not get
removed from the local store. However, it does get the 'isError'
flag set to 'true'. So we now filter out any resources that we know
have been deleted.
Also fixed some computed properties in the related-list component.
TODO: fix component logic for creating relations