Fixed dirty property checking for relationship attributes
I discovered that Ember does not properly track canonical state and
current state for 'hasMany' relationships (e.g. array attributes)
and 'belongsTo' relationships, depending upon if the newly created
object was saved without reloading the page, or if the resulting
attribute is blank/empty. Part of the reason is that Ember expects
each 'object attribute' to have an ID and properties explicitly
stated, but FHIR will remove empty object attributes and arrays.
This causes Ember not to update canonical state correctly when
resources are reloaded.
So, I needed to implement a workaround to make sure canonical and
current state are properly synchronized when objects are reloaded.

depending upon if the newly created object was